+++ x/pylons/keeper/msg_server_fulfill_trade_test.go
@@ -1,3 +1,45 @@
 package keeper_test
 
-// TODO add tests
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/Pylons-tech/pylons/x/pylons/keeper"
+	"github.com/Pylons-tech/pylons/x/pylons/types"
+)
+
+func (suite *IntegrationTestSuite) TestFulfillTradeMsgServerSimple() {
+	k := suite.k
+	ctx := suite.ctx
+	require := suite.Require()
+
+	wctx := sdk.WrapSDKContext(ctx)
+	srv := keeper.NewMsgServerImpl(k)
+
+	creatorA := types.GenTestBech32FromString("creatorA")
+	creatorB := types.GenTestBech32FromString("creatorB")
+
+	for i := 0; i < 5; i++ {
+		msgCreate := &types.MsgCreateTrade{
+			Creator:     creatorA,
+			CoinInputs:  nil,
+			ItemInputs:  nil,
+			CoinOutputs: nil,
+			ItemOutputs: nil,
+			ExtraInfo:   "extrainfo",
+		}
+
+		respCreate, err := srv.CreateTrade(wctx, msgCreate)
+		require.NoError(err)
+		require.Equal(i, int(respCreate.ID))
+
+		msgFulfill := &types.MsgFulfillTrade{
+			Creator:         creatorB,
+			ID:              respCreate.ID,
+			CoinInputsIndex: 0,
+			Items:           nil,
+		}
+
+		_, err = srv.FulfillTrade(wctx, msgFulfill)
+		require.NoError(err)
+	}
+}